requirements.txt
pyserial~=3.5
SQLAlchemy~=1.4.29